pyproject.toml
[project]
name = "reelcraft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "firecrawl-py>=4.5.0",
    "google-genai>=1.46.0",
    "httpx>=0.28.1",
    "langfuse>=3.8.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",
    "boto3>=1.40.59",
]
